requirements.txt
streamlit>=1.28.0
openai>=1.0.0
groq>=0.4.0
google-generativeai>=0.3.0
tavily-python>=0.3.0
google-api-python-client>=2.100.0
requests>=2.28.0
sentence-transformers>=2.6.0
python-dotenv>=1.0.0
pypdf>=4.0.0
python-docx>=1.0.0
